Changes to support custom installation and uninstallation. UIpdated to last Squirrel.windows Also refactorized Application Path (WinRAssemblyInfo.DefaultPath)

+++ projects/WinR.Core/Installation/UninstallWinRShortCuts.cs
@@ -9,10 +9,13 @@ namespace WinR.Core.Installation
         internal void Execute()
         {
             var sendToFilePath = Path.Combine(Environment.GetFolderPath(Environment.SpecialFolder.SendTo), WinRAssemblyInfo.Product + ".lnk");
-            var desktopFilePath = Path.Combine(Environment.GetFolderPath(Environment.SpecialFolder.CommonDesktopDirectory), WinRAssemblyInfo.Product + ".lnk");
-            
-            File.Delete(sendToFilePath);
-            File.Delete(desktopFilePath);
+            var desktopFilePath = Path.Combine(Environment.GetFolderPath(Environment.SpecialFolder.DesktopDirectory), WinRAssemblyInfo.Product + ".lnk");
+
+            if (File.Exists(sendToFilePath))
+                File.Delete(sendToFilePath);
+
+            if (File.Exists(desktopFilePath))
+                File.Delete(desktopFilePath);
         }
     }
 }
